% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aclust.listDemo.R
\docType{data}
\name{aclust.listDemo}
\alias{aclust.listDemo}
\title{This is data to be included in my package}
\format{A list with 500 slots}
\usage{
data(aclust.listDemo)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
