% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2M_wrapper.R
\name{beta2M_wrapper}
\alias{beta2M_wrapper}
\title{wrapper function to do beta to m value transformation}
\usage{
beta2M_wrapper(beta2M = TRUE, BetaPhenoTrain_df, BetaPhenoTest_df,
  respCol_index, returnType = "matrix")
}
\arguments{
\item{beta2M}{whether transfre beta to m value before prediction}

\item{BetaPhenoTrain_df}{train data frame that each column represents a CpG probe
while each row represents a sample, each cell is a M value, first column
is phenodata}

\item{BetaPhenoTest_df}{test data frame that each column represents a CpG probe
while each row represents a sample, each cell is a M value, first column
is phenodata}

\item{respCol_index}{response variable col number in beta data frame}

\item{returnType}{return data type, data frame or matrix, default set to matrix}
}
\value{
A list of two elements: train data frame and test data frame
}
\description{
wrapper function to do beta to m value transformation
}
\examples{
\dontrun{
   data(ExampleMvalue_train)
   data(ExampleMvalue_test)

   test <- beta2M_wrapper(
     beta2M = TRUE,
     BetaPhenoTrain_df = ExampleMvalue_train,
     BetaPhenoTest_df = ExampleMvalue_test,
     respCol_index = 1,
     returnType = "matrix"
   )
 }
}
