% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretWrapper.R
\name{caretWrapper}
\alias{caretWrapper}
\title{warpper function of parameters for ramdom forest or suppport vector machine}
\usage{
caretWrapper(trainTable_df, respCol_index, predMethod = c("rf",
  "svmLinear"), seed_int, whichCVfold_int, ncores = 2)
}
\arguments{
\item{trainTable_df}{train data frame that each column represents a CpG probe
while each row represents a sample, each cell is a methaylation M value}

\item{respCol_index}{response variable col number in beta data frame}

\item{predMethod}{use random forest or svm to do prediction}

\item{seed_int}{seed.value in use right row}

\item{whichCVfold_int}{crossvalidation num in use right now, this decides on the traning data to be used}

\item{ncores}{number of cores to do parallel computing}
}
\value{
A list contains results of glmnet
}
\description{
warpper function of parameters for ramdom forest or suppport vector machine
}
\examples{
\dontrun{
 data(ExampleMvalue_train)
 fit <- caretWrapper(trainTable_df = ExampleMvalue_train,
               respCol_index = 1,
               predMethod = "rf",
               seed_int = 120,
               whichCVfold_int = 5,
               ncores = 2)
 }
}
