% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chrome_annot_data.R
\docType{data}
\name{chrome_annot_files}
\alias{chrome_annot_files}
\title{This is data to be included in my package}
\format{A list with 22 elements}
\usage{
data(chrome_annot_files)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
