% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateGlmnet.R
\name{evaluateGlmnet}
\alias{evaluateGlmnet}
\title{function to do prediction and evaluate based on the results of glmnet}
\usage{
evaluateGlmnet(fit, alpha, testTable, respCol_index, whichRep_int,
  whichCVfold_int)
}
\arguments{
\item{fit}{fit result of glmnet}

\item{alpha}{specific alpha value use for this glmnet fit object}

\item{testTable}{test data frame that each column represents a CpG probe
while each row represents a sample, each cell is a M value, first column
is phenodata}

\item{respCol_index}{col number of response variable}

\item{whichRep_int}{number of repetation}

\item{whichCVfold_int}{number of cross validation}
}
\value{
A data frame contains parameters for evaluate prediction performance
  \itemize{
    \item{\code{NumOfRep} : }{Number of repetation}
    \item{\code{NumOfCv} : }{Number of cross validation}
    \item{\code{alphaValue} : }{alpha value of glmnet}
    \item{\code{lambda.min} : }{lambda min value of glmnet}
    \item{\code{auc_results} : }{auc value of prediction}
    \item{\code{Sensitivity} : }{Sensitivity value of prediction}
    \item{\code{Specificity} : }{Specificity value of prediction}
    \item{\code{...} : }{more index}
  }
}
\description{
function to do prediction and evaluate based on the results of glmnet
}
\details{
evaluate performance of results return by function \link{glmnetWrapper}
}
\examples{
\dontrun{
   data(ExampleMvalue_test)
   data(glmnet_Fit_ls)
   alphaValue = seq(0, 1, by = 0.1)

   test <- evaluateGlmnet(
     fit = glmnet_Fit_ls[[1]],
     alpha = alphaValue[1],
     testTable = ExampleMvalue_test,
     respCol_index = 1,
     whichRep_int = 1,
     whichCVfold_int = 1
     )
 }
}
