% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateRF.R
\name{evaluateRF}
\alias{evaluateRF}
\title{function to do prediction and evaluate based on the results of random forest}
\usage{
evaluateRF(caret_fit, testTable, respCol_index, whichRep_int,
  whichCVfold_int)
}
\arguments{
\item{caret_fit}{results of random forest or support vector machine}

\item{testTable}{test data frame that each column represents a CpG probe
while each row represents a sample, each cell is a M value, first column
is phenodata}

\item{respCol_index}{col number of response variable}

\item{whichRep_int}{repetition num in use right now, this decides on the
traning data to be used}

\item{whichCVfold_int}{crossvalidation num in use right now, this decides
on the traning data to be used}
}
\value{
A data frame of containing the train and test beta matrix
  \itemize{
    \item{\code{auc_results} : }{auc value of prediction}
    \item{\code{Sensitivity} : }{Sensitivity value of prediction}
    \item{\code{Specificity} : }{Specificity value of prediction}
    \item{\code{...} : }{more index}
  }
}
\description{
function to do prediction and evaluate based on the results of random forest
}
\examples{
\dontrun{
   data(ExampleMvalue_test)
   data(rf_Fit)

   test <- evaluateRF(
     caret_fit = rf_Fit,
     testTable = ExampleMvalue_test,
     respCol_index = 1,
     whichRep_int = 1,
     whichCVfold_int = 1
   )
 }
}
