% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examplebeta_df.R
\docType{data}
\name{Example_df}
\alias{Example_df}
\title{This is data to be included in my package}
\format{A data frame with 23501 rows and 110 variables}
\usage{
data(Example_df)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
