% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ExampleMvalue_df.R
\docType{data}
\name{ExampleMvalue}
\alias{ExampleMvalue}
\title{This is data to be included in my package}
\format{A data frame with 110 rows and 23502 variables}
\usage{
data(ExampleMvalue)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
