% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ExampleMvalue_test_df.R
\docType{data}
\name{ExampleMvalue_test}
\alias{ExampleMvalue_test}
\title{This is data to be included in my package}
\format{A data frame with 22 rows and 23502 variables}
\usage{
data(ExampleMvalue_test)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
