% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ExampleMvalue_train_df.R
\docType{data}
\name{ExampleMvalue_train}
\alias{ExampleMvalue_train}
\title{This is data to be included in my package}
\format{A data frame with 88 rows and 23502 variables}
\usage{
data(ExampleMvalue_train)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
