% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPC1.R
\name{getPC1}
\alias{getPC1}
\title{get PC1 of cpg list}
\usage{
getPC1(clust_ls, train_df, test_df)
}
\arguments{
\item{clust_ls}{list of cpgs}

\item{train_df}{beta or M value matrix with train samples}

\item{test_df}{beta or M value matrix with test samples}
}
\value{
a list contains performance of predicition
}
\description{
given a beta matrix with list of cpgs, extract pc1 of each slot
   of the list and return the pc1 score data frame
}
\examples{
\dontrun{
  data(aclust.listDemo)
  data(ExampleMvalue_train)
  data(ExampleMvalue_test)

   test <- getPC1(
   clust_ls = aclust.listDemo,
   train_df = ExampleMvalue_train[,-1],
   test_df = ExampleMvalue_test[,-1]
 )
}
}
