% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glmnet_Fit_ls.R
\docType{data}
\name{glmnet_Fit_ls}
\alias{glmnet_Fit_ls}
\title{This is data to be included in my package}
\format{A list of glmnet results}
\usage{
data(glmnet_Fit_ls)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
