% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexMeasure.R
\name{indexMeasure}
\alias{indexMeasure}
\title{function that input a data table, returns a table that contains
mean_cross_validation with alpha-lambda value that makes the index maxium
methods represents specific methods you choose}
\usage{
indexMeasure(data_table, methods = c("CpGs_within_cluster",
  "CpGs_most_variablity", "CpGs_most_significant"),
  index = c("Balanced.Accuracy", "Sensitivity", "Specificity", "auc",
  "F1"))
}
\arguments{
\item{data_table}{data table that}

\item{methods}{method that we care about}

\item{index}{index that we want to see the performance}
}
\description{
function that input a data table, returns a table that contains
mean_cross_validation with alpha-lambda value that makes the index maxium
methods represents specific methods you choose
}
\examples{
\dontrun{
 phenoSplit(data_table = a[[1]],
            methods = "CpGs_most_variablity",
            index = "Sensitivity")
}
}
