% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logitB.R
\name{logitB}
\alias{logitB}
\title{function that do logit transformation}
\usage{
logitB(x, B = exp(1))
}
\arguments{
\item{x}{a numeric or complex vector.}

\item{B}{a positive or complex number: the base with respect to which
logarithms are computed. Defaults to e=exp(1)}
}
\value{
value after logit transformation
}
\description{
function that do logit transformation
}
\examples{
\dontrun{
  logitB(0.5, B = 2)

}
}
