% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methSplit.R
\name{methSplit}
\alias{methSplit}
\title{function that split methylation data into train and test}
\usage{
methSplit(rowNum, Beta_df, designInfo_df)
}
\arguments{
\item{rowNum}{number of row of information_df}

\item{Beta_df}{data frame that each row is a cpg probe, each col is a
sample id, each cell is a M value, first column is the phenodata}

\item{designInfo_df}{df create by function \link{summaryInfo}}
}
\value{
A list of two elements containing the train and test beta matrix
  \itemize{
    \item{\code{Train_df} : }{A data frame of the train sample beta values
      }
    \item{\code{Test_df} : }{A data frame of the test sample beta values
      }
  }
}
\description{
function that split methylation data into train and test
}
\examples{
\dontrun{
 data(Example_df)
 data(pfcInfo_df)
 data_ls <- methSplit(
   rowNum = 1,
   Beta_df = Example_df,
   designInfo_df = pfcInfo_df
 )
}
}
