% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parTtest.R
\name{parTtest}
\alias{parTtest}
\title{function that do parallel t test and elect results based on p.threshold}
\usage{
parTtest(group1, group2, ncores = 2)
}
\arguments{
\item{group1}{df that each col is a probe id, each row is a sample}

\item{group2}{df that each col is a probe id, each row is a sample}

\item{ncores}{number of cores to do parallel computing}
}
\description{
function that do parallel t test and elect results based on p.threshold
}
\examples{
\dontrun{
 parTtest(earlygroup, lategroup, ncores = 2)
}
}
