% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pfcInfo_df.R
\docType{data}
\name{pfcInfo_df}
\alias{pfcInfo_df}
\title{This is data to be included in my package}
\format{A data frame with 50 rows and 113 variables}
\usage{
data(pfcInfo_df)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
