% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pfc_pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{This is data to be included in my package}
\format{A data frame with 110 rows and 4 variables}
\usage{
data(pheno)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
