% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wrapper.R
\name{predict_wrapper}
\alias{predict_wrapper}
\title{wrapper function of three prediction methods}
\usage{
predict_wrapper(predictMethod = c("glmnet", "randomForest", "svm"),
  alphaValue, PhenoTrain_df, PhenoTest_df, respCol_index, outcome_type,
  seed_int, whichRep_int, whichCVfold_int, ncores)
}
\arguments{
\item{predictMethod}{what prediction method to use}

\item{alphaValue}{vector that storage alpha values}

\item{PhenoTrain_df}{data frame that each row is a train samples}

\item{PhenoTest_df}{data frame that each row is a test samples}

\item{respCol_index}{response variable col number in beta data frame}

\item{outcome_type}{type of outcome variable, gauusian or binomial or poisson, etc}

\item{seed_int}{seed.value in use right row}

\item{whichRep_int}{repetition num in use right now, this decides on the traning data to be used}

\item{whichCVfold_int}{crossvalidation num in use right now, this decides on the traning data to be used}

\item{ncores}{number of cores to do parallel computing}
}
\value{
a list contains train data frame and test data frame and number of predictors
}
\description{
wrapper function of three prediction methods
}
\examples{
\dontrun{
 data(ExampleMvalue_train)
 data(ExampleMvalue_test)

 predict_ls <- predict_wrapper(
   predictMethod = "glmnet",
   alphaValue = seq(0, 1, by = 0.1),
   PhenoTrain_df = ExampleMvalue_train,
   PhenoTest_df = ExampleMvalue_test,
   respCol_index = 1,
   outcome_type = "binomial",
   seed_int = 123,
   whichRep_int = 1,
   whichCVfold_int = 1,
   ncores = 2
   )
 }
}
