% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rf_Fit.R
\docType{data}
\name{rf_Fit}
\alias{rf_Fit}
\title{This is data to be included in my package}
\format{A list with results of random forest}
\usage{
data(rf_Fit)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
