% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAcluster.R
\name{runAcluster}
\alias{runAcluster}
\title{function that run parallel acluster algorithm given a methylation data frame}
\usage{
runAcluster(chromeAnnot_df, tBeta_df, minCpGs = 5, ncores = 2)
}
\arguments{
\item{chromeAnnot_df}{data frame contains 1 chrome annot information for 450k probes}

\item{tBeta_df}{a long format BetaValue_df to generate aclusters}

\item{minCpGs}{minimum number of cpgs within each cluster}

\item{ncores}{number of cores to do parallel computing}
}
\value{
A list of elements containing cg probe names
}
\description{
function that run parallel acluster algorithm given a methylation data frame
}
\examples{
\dontrun{
 data(Example_df)
 data(chrome_annot_files)
 BetaTrans_df <- TransposeAssay(
 Example_df[, -1],
 omeNames = "rowNames"
 )
 runAcluster(
    chromeAnnot_df = chrome_annot_files,
    tBeta_df = BetaTrans_df,
    minCpGs = 5,
    ncores = 2
 )
}

}
