% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParameters.R
\name{setParameters}
\alias{setParameters}
\title{function that set parameters for glmnet}
\usage{
setParameters(nAlpha, nLambda)
}
\arguments{
\item{nAlpha}{how many number of alpha values you want}

\item{nLambda}{how many number of nLambda values you want}
}
\value{
A list of two elements containing the alpha and lambda value
  \itemize{
    \item{\code{alpha} : }{alpha value for glmnet
      }
    \item{\code{lambda} : }{lambda value for glmnet
      }
  }
}
\description{
function that set parameters for glmnet
}
\examples{
\dontrun{
 parameters_ls <- setParameters(nAlpha = 11,
                               nLambda = 11)
}
}
