% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryInfo.R
\name{summaryInfo}
\alias{summaryInfo}
\title{function that create a information data frame that storage all parameters we need}
\usage{
summaryInfo(NumOfRep, NumOfCv, seed, phenoData_df, respCol_index)
}
\arguments{
\item{NumOfRep}{number of repeat measurements in this trial}

\item{NumOfCv}{number of cross validation in this trail}

\item{seed}{seed value}

\item{phenoData_df}{a data frame contains sample id(in first column) and variables
we interested in(in the second column), the first column should named "Sample"}

\item{respCol_index}{response variable col number in beta data frame}
}
\value{
A data frame containing design experiment information
  \itemize{
    \item{\code{NumOfRep} : }{number of repeat measurements in this trial
      }
    \item{\code{NumOfCv} : }{number of cross validation in this trail
      }
    \item{\code{seed} : }{seed value for trail
      }
    \item{\code{...} : }{sample names
      }
  }
}
\description{
function that create a information data frame that storage all parameters we need
}
\examples{
\dontrun{
 data(pheno)
 information_df <- summaryInfo(
  NumOfRep = 10,
  NumOfCv = 5,
  seed = 122,
  phenoData_df = pheno,
  respCol_index = 2
 )
}
}
