% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_svm_Fit.R
\docType{data}
\name{svm_Fit}
\alias{svm_Fit}
\title{This is data to be included in my package}
\format{A list with results of svm}
\usage{
data(svm_Fit)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
